Agent Web/requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-core==0.1.32
google-generativeai==0.3.2
python-dotenv==1.0.1
duckduckgo-search==4.4.3
langchain-community==0.0.27
langchain-google-genai==0.0.11 